+++ trionyx/quickstart/project/config/__init__.py
@@ -2,3 +2,4 @@
 from celery_app import app as celery_app
 
 __all__ = ['celery_app']
+__version__ = '0.0.1'