+++ trionyx/quickstart/project/config/settings/live.py
@@ -1,16 +1,11 @@
 """Live project settings"""
 from .base import *  # noqa: F401,F403
+import os
 
-DATABASES = {
-    "default": {
-        "ENGINE": "django.db.backends.postgresql",
-        "NAME": "",
-        "USER": "",
-        "PASSWORD": "",
-        "HOST": "",
-        "PORT": "",
-    }
-}
+
+LOG_DIR = get_env_var('LOG_DIR', BASE_DIR)
+os.makedirs(LOG_DIR, exist_ok=True)
+
 
 LOGGING = {
     'version': 1,
@@ -24,22 +19,20 @@ LOGGING = {
     'handlers': {
         'error-file': {
             'level': 'WARNING',
-            'class': 'logging.FileHandler',
-            'filename': os.path.join(BASE_DIR, 'error.log'),
+            'class': 'logging.handlers.RotatingFileHandler',
+            'filename': os.path.join(LOG_DIR, 'error.log'),
+            'maxBytes': 1024*1024*5, # 5 MB
+            'backupCount': 10,
             'formatter': 'verbose'
         },
     },
     'loggers': {
         'apps': {
             'level': 'WARNING',
-            'handlers': ['file', 'error-file'],
+            'handlers': ['error-file'],
         },
         'trionyx': {
             'level': 'WARNING',
-            'handlers': ['error-file'],
-        },
-        'django_jsend': {
-            'level': 'ERROR',
             'handlers': ['error-file'],
         },
         'django': {