+++ trionyx/quickstart/reusable_app/[app_name]/__init__.py
@@ -1,2 +1,4 @@
 """App version"""
 __version__ = '0.0.1'
+
+default_app_config = '[[app]].apps.Config'