+++ trionyx/trionyx/views.py
@@ -5,11 +5,15 @@ trionyx.trionyx.view.accounts
 :copyright: 2018 by Maikel Martens
 :license: GPLv3
 """
+import os
+import re
 import json
 import logging
 from collections import OrderedDict
 
-from django.http import HttpResponse, HttpResponseRedirect
+from docutils.core import publish_parts
+from django.core.paginator import Paginator
+from django.http import HttpResponse, HttpResponseRedirect, JsonResponse
 from django.contrib.auth.views import LoginView as DjangoLoginView
 from django.contrib.auth import logout as django_logout
 from django.contrib.auth.models import Permission
@@ -22,6 +26,7 @@ from django.conf import settings
 from django.contrib.staticfiles.templatetags.staticfiles import static
 from django.utils.translation import ugettext_lazy as _
 from django.contrib import messages
+from django.forms.widgets import SelectMultiple
 
 from trionyx.models import get_class
 from trionyx.views import UpdateView, DetailTabView, DialogView, JsendView
@@ -30,7 +35,7 @@ from trionyx.config import models_config
 from trionyx.widgets import widgets
 from trionyx import utils
 from trionyx.forms.helper import FormHelper
-from trionyx.forms import form_register, modelform_factory
+from trionyx.forms import form_register, modelform_factory, ModelAjaxChoiceField
 from trionyx.models import filter_queryset_with_user_filters
 from trionyx.trionyx.tasks import MassUpdateTask
 
@@ -38,6 +43,15 @@ User = get_class('trionyx.User')
 Task = get_class('trionyx.Task')
 
 logger = logging.getLogger(__name__)
+
+
+def basic_auth(request, user_type):
+    """View that can be used for basic-auth like Nginx ngx_http_auth_request_module"""
+    if user_type == 'superuser' and request.user.is_authenticated and request.user.is_superuser:
+        return HttpResponse('OK')
+    if user_type == 'user' and request.user.is_authenticated:
+        return HttpResponse('OK')
+    return HttpResponse('NOK', status=403)
 
 
 def media_nginx_accel(request, path):
@@ -197,6 +211,43 @@ def create_permission_jstree(selected=None, disabled=False):
 
 
 # =============================================================================
+# Ajax form views
+# =============================================================================
+def ajaxFormModelChoices(request, id):
+    """View for select2 ajax data request"""
+    if id not in ModelAjaxChoiceField.registered_fields:
+        return JsonResponse({})
+
+    field = ModelAjaxChoiceField.registered_fields[id]
+    query = field.queryset.order_by('verbose_name')
+
+    search = request.GET.get('q', None)
+    if search:
+        config = models_config.get_config(query.model)
+        if not config.disable_search_index:
+            query = watson.filter(query, search, ranking=False)
+        else:
+            query = query.filter(verbose_name__contains=search)
+
+    pages = Paginator(query, 20)
+    page = pages.page(int(request.GET.get('page', 1)))
+
+    result = [
+        {'id': '', 'text': '------'}
+    ] if not isinstance(field.widget, SelectMultiple) and not field.required and page.number == 1 else []
+
+    return JsonResponse({
+        'results': result + [{
+            'id': row[0],
+            'text': row[1],
+        } for row in page.object_list.values_list('id', 'verbose_name')],
+        'pagination': {
+            'more': page.has_next(),
+        }
+    })
+
+
+# =============================================================================
 # Global search
 # =============================================================================
 class GlobalSearchJsendView(JsendView):
@@ -300,6 +351,40 @@ class UserTasksJsend(JsendView):
                 'url': task.get_absolute_url(),
             } for task in tasks
         ]
+
+
+# =============================================================================
+# Changelog
+# =============================================================================
+class ChangelogDialog(DialogView):
+    """Dialog to show app changelog based on CHANGELOG.rst"""
+
+    def display_dialog(self):
+        """Display changelog"""
+        changelog_path = os.path.join(settings.BASE_DIR, 'CHANGELOG.rst')
+
+        if os.path.isfile(changelog_path):
+            with open(changelog_path, 'r', encoding='utf-8') as _file:
+                content = publish_parts(_file.read(), writer_name='html')['html_body']
+        else:
+            content = ''
+
+        if settings.TX_CHANGELOG_HASHTAG_URL:
+            link_re = r'<a href="{link}" target="_blank">#\1</a>'.format(link=settings.TX_CHANGELOG_HASHTAG_URL.format(tag='\\1'))
+            content = re.sub(r"#([\d\w\-]+)", link_re, content)
+
+        return {
+            'title': str(_('Changelog for {name}')).format(name=settings.TX_APP_NAME),
+            'content': f'<div class="changelog-wrapper">{content}</div>',
+            'submit_label': _("Don't show again") if self.request.GET.get('show') else False,
+        }
+
+    def handle_dialog(self):
+        """Save shown changelog version"""
+        self.request.user.set_attribute('trionyx_last_shown_version', utils.get_app_version())
+        return {
+            'close': True,
+        }
 
 
 # =============================================================================
@@ -417,7 +502,7 @@ class WidgetConfigDialog(DialogView):
                 config = form.cleaned_data
                 config['title'] = self.request.POST.get('title')
                 config['refresh'] = self.request.POST.get('refresh')
-            else:
+            elif '__post__' in self.request.POST:
                 logger.error(json.dumps(form.errors))
         elif '__post__' in self.request.POST:
             config = {